Organize service utilities
