+++ src/services/suggest_character_roles.py
@@ -6,7 +6,7 @@ import json
 import re
 from typing import Any, Dict, List
 
-from openai_service import client
+from .openai_service import client
 from logging_config import get_logger
 
 logger = get_logger(__name__)
